[roll] Roll fuchsia [omaha-client] Reboot sooner if a user asks for an UpdateCheck
If a user initiates an update check in the middle of an update check,
or while we're waiting for a reboot, treat the original update check
as if it was user-initiated. This "updgrades" all update checks to act
as if they were user-initiated if a user asks for a check while
they're running.

Test: added a StateMachine test for a user-initiated update check
while waiting for reboot, and a test for a user-initiated update check
while applying an update (both cases in which we can upgrade in flight
updates).

MULTIPLY: omaha-client-lib-tests

Original-Fixed: 62821
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444925
Original-Revision: 0275fcecc20f05ab91b8a599738ecc84e493c0af
GitOrigin-RevId: 663123fcd02847c2864dc1ed506e099c149147fc
Change-Id: Ifb0a4994363629ac6b7742f81396025aff28f1ab