[roll] Roll fuchsia [kernel][sched] Fix precedence edge in thread migration.
Fix an edge case in the precedence of thread migration and time
slice expiration conditions. When an action occurs that requires
thread migration, such as setting a thread's affinity to exclude
the CPU it is running on, a reschedule is required to carry out
the migration. If the expiration of the time slice of the thread
coincides with the need to migrate, the reschedule will first
return the thread to the run queue of the current CPU, when it
should instead migrate immediately to a different CPU.

The kernel test trylock_or_cancel_canceled (k ut timer) is an
example of a situation that can result in a deadlock because of
the thread returning to the run queue instead of migrating as soon
as possible.

Original-Bug: 62970
Test: Multiple instances of QEMU+KVM running k ut -r 500000 timer.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445815
Original-Revision: bd329a0b8563c7cb73c8ee54aaa6b9b4bb30f8cb
GitOrigin-RevId: d98ea35ac970a706a421f51e302b823e19bf5c1f
Change-Id: I5ff98321d2c6d70ad1e036007e8876b1da9112b0