[roll] Roll fuchsia [kernel][vm] Guard VmCowpages::options_ with lock
The options_ member is currently always accessed with the lock being
held so this change is semantically a no-op, although some code
juggling is needed to convince the static analysis.

Motivation for this is that later on we may want to have nodes
transition to and from different states (such as hidden and not hidden)
and so by guarding options_ now we ensure that no unlocked usages are
added in the future.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437349
Original-Revision: e266166fd3fb67c6d9efa787dcf5a8e965420166
GitOrigin-RevId: 3bd63b28e03898ff588f5cdfd438ec6b9417a672
Change-Id: I817f2dbe125f3ab42c78d18a0197fa89d242f2b1