[roll] Roll infra/recipes [build] Fix compdb infra failures
If a command in the compdb uses an executable from a directory like
`/bin` that's not in one of the known recipe directories like
`api.path["start_dir"]`, the recipe would fail trying to convert the
executable's path to a `Path` object using `api.path.abs_to_path()`. We
only do this to figure out whether the path is in the build directory,
so if `abs_to_path` fails then the path is definitely not in the build
directory, and we can skip it.

This was breaking the firebase-docs builder.

Original-Fixed: 63216
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/445477
Original-Revision: c4571962f65c5c3406278581f413653ac540f4eb
GitOrigin-RevId: 31873a8e9be5a692d373ebb9ea7549b0b0231e07
Change-Id: Ida32325a643719e4a64df167893526212a6ce72b