[roll] Roll fuchsia [sfw] Relay InputDeviceRegistry requests to Session
The `fuchsia.input.injection.InputDeviceRegistry` protocol
is implemented by a Session (which is a V2 component),
but needs to be accessible to V1 components.

Normally, V2 components are made available to V1 components
using capability routing. However, that only works for
static children of a component, and the Session is
a dynamic child of the session manager.

Thus, to make `InputDeviceRegistry` available to V1 components
(such as the `input` command-line tool and `sl4f`), the session
manager relays `InputDeviceRegistry` requests to the Session.

Note that `InputDeviceRegistry` protocol is used only to connect
fake devices to the input pipeline, _not_ to inject input
events into the pipeline. Hence, this relay does not introduce
latency on the data path.

Original-Bug: 62363
Test: fx test session_manager_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445309
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>
Original-Revision: acbe1f7d5740652b7edd8b3187205e22bd5171c9
GitOrigin-RevId: 53f9bf22cbd685b5d79b66ade5f77a584b7eb132
Change-Id: I446c27a81964a8647a455c47beec26c70f0fde38