[roll] Roll fuchsia [kernel][lib/arch] Rename test files
The "x86-" prefix does not add any real information, and lends itself to
an eyesore of a naming scheme as more tests are added. Moreover, we
qualify the existing CPUID tests as "basic", to make room for other
test files (soon to be added) that will introduce other CPUID-related
content.

We do this now, so as not to complicate subsequent changes.

Original-Bug: 61093
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446414
Original-Revision: 48c43d409924467384c35006e727e72a63a5bd63
GitOrigin-RevId: 8c9d78e498ea7a53ba393b004fa1679e7e444adc
Change-Id: Ic1beffc6f3382c9f6d3ab1e94c657de8bef3e692