[roll] Roll fuchsia [wlan][policy] remember active scan connections
This change will remember when a network config was used to connect to a
network after the network required an active scan to be seen. This
reduces reconnect time for hidden networks and increases the probability
of seeing a hidden network in a scan.

Since we can now have conflicting data indicating hidden network or not
hidden network (ie seen in passive scan and connected to after active
scan), we will use whatever has happened most recently.

HiddenStats is removed for now because this calculation of hidden
probability does not actually need that information. It will be added
back if/when it is needed.

Multiply: wlancfg-tests
Test: Modified and added tests:
	fx test wlancfg-tests
Original-Bug: 62646
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445556
Original-Revision: bb718840c672aab572544c2ffcb058786108f900
GitOrigin-RevId: 01847d673f6b11c1c7aaaaf7734ac3a55c9044f5
Change-Id: Ic2d040ab8c48c6a5f044af08681b0d933e083f2d