[roll] Roll fuchsia [audio][lib] Add Buffer::Append, fix Slice::begin|end
This CL adds the ability to add an AudioBuffer or AudioBufferSlice
to the end of an AudioBuffer. Implementation was blocked by a bug
in AudioBufferSlice::begin() and end(), which are now also fixed.
Unittests are added for the Append methods, and existing unittests
are expanded (they now fail without the fixes).

Test: AudioBuffer and AudioBufferSlice are test-only

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446550
Original-Revision: cadbd18f0db123bbe6920aa849a65ba12af805d8
GitOrigin-RevId: 9e9d1ef45d0e2b5ce2492c98ee264063d0267702
Change-Id: I2c51510641da239b001b2b4bf87390b6d12d0a56