[roll] Roll fuchsia [debugger] Fix crash on build ID error.
The symbol server code would crash attempting to extract an invalid
substring if the build ID was empty or very short. This change avoids
the crash and reports an error. This does not solve the problem where
build IDs are wrong in some cases.

Fixes some cases in the symbol server code where an error would be
reported but the corresponding Err() message was unset. Clean up some
return values and use the ErrOr template to make it more difficult to
make this kind of mistake.

Original-Bug: 63298
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446595
Original-Revision: b7b6a0356019a353da898b36725d51f8388675b9
GitOrigin-RevId: d36049eb548f6daad74beee94fbeba3674256c53
Change-Id: I7a467375f413155c3bf651c44eb3a93dc2d49c9a