[roll] Roll fuchsia [system-updater] Simplify boot metadata changes.
Before writing updates, the system-updater needs to put boot metadata
into a happy state. The happy state is "current is active and
alternate is unbootable". Before this change, the system-updater took
a number of steps that depended on whether current was already active,
and whether current was marked successful or not.

This change simplifies that substantially. The new algorithm is:

1. If the current configuration is not marked successful, abort the
   update.
2. Mark the alternate configuration unbootable.

Put together, this ensures that current is both active and successful,
and that alternate is unbootable. The "cost" of this simplicitly is
that the update fails if current isn't marked as successful, but
that's aligned with our stance than unhealthy systems shouldn't OTA.

This change deletes a lot of code and tests, largely because the new
algorithm is completely agnostic to which configuration is "active" at
the start of the update.

Original-Bug: 62958
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441514
Original-Revision: f678e905ea2f9c91f14bc43d6717c5da200989cf
GitOrigin-RevId: 833da6bacf2422bf84c6449905606368bba62695
Change-Id: I02abd583674df1f3acf399ae706da02c7197f157