[roll] Roll fuchsia [system-health-check] Make system-health-check more testable.
Currently, system_health_check::set_active_configuration_healthy gets
a handle to the PaverService via the envionrment, which is why we have
a whole complicated integration test that wraps the library in a
binary just to pass the mock paver in.

Instead, set_active_configuration_healthy now accepts a
BootManagerProxy so it's much easier to test, and no longer has
implicit dependencies on the environment. Accordingly, this also
changes system-update-checker and omaha-client to explicitly provision
the paver for the system-health-check library.

All the test that were in mark-active-configuration-healthy-tests are
now unit tests.

Context: while I think this is a positive change regardless, I've
mostly put it into its own CL in order to reduce the line count of a
much bigger CL to fix fxb/51480 (see next in chain).

Original-Bug: 51480

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442995
Original-Revision: 1b0a7f71a5c429bfe39702596697b097639517c7
GitOrigin-RevId: 26df237f1454f5a3f203058f2d15db18d7d5c322
Change-Id: I21b145b8418ccb4c02af87dc721f06ab50f5beee