[roll] Roll fuchsia [bt][rfcomm] Cancel pending channel requests if establishment fails
If the Session terminates, or a specific DLC is not able to be
established, notify clients of the cancellation. This addresses the
issue of a profile client requesting an RFCOMM connection (using
Profile.Connect), the peer either disconnecting or rejecting the
request, and the connection request remaining hanging. In such cases, we
now notify the client immediately of the failure.

Original-Bug: 59585
Test: Added to bt-rfcomm-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/443861
Original-Revision: d9ec797063874c88551effbe6b190ea5a837039e
GitOrigin-RevId: 1227ea03324f140aa7855aa291df44b51ca43663
Change-Id: I4a7788fc6a25dfa6622b589796a448f77b276d75