[roll] Roll fuchsia [oom][test] Disable flaky oom_test
The oom_test is flaking because a crash in driver_manager is causing the
device to reboot early. The test can sometimes trigger the OOM too soon,
catching the driver_manager in the middle of initialization, such that
allocations start failing with ZX_ERR_NO_MEMORY. This causes the
driver_manager to crash, and since it is a critical process, the device
reboots. This fails the test which specifically watches for an OOM
reboot.

Disable the test until it is rewritten to be more reliable.

Original-Bug: 63505
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447296
Original-Revision: 10a7abb024c58f386b63b60a005a2e94f0509e8e
GitOrigin-RevId: a584da06acbf2cdcca561879c5d245ccbeb922f0
Change-Id: I3582d7d0596db70a98624b53598c51909b683225