[roll] Roll fuchsia [display] Fake fuchsia.hardware.display.Provider service.
Refactors display driver testing code so that it can be repurposed
in an implementation of fuchsia.hardware.display.Provider which
vends a hermetic fake display controller, or rather: mostly hermetic.
This is because it uses the "global" sysmem.Allocator obtained from
the environment; this is so that E2E tests can be configured to
share the same Sysmem between the display controller, Vulkan,
and Scenic.

TEST=fx test display-core-unittests
TEST=fx test flatland_renderer_unittests
TEST=fx test root_presenter_unittests

Original-Bug: 62388

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427501
Original-Revision: 3c4cf68e93e04e51deceb263a27a4c0e49dc21e8
GitOrigin-RevId: 6c8f12a7cd4ae075b033be79f056ea94a91f6ebb
Change-Id: I6a0c1ea6df9ebe565fb3ec0c965bb3b467a9f014