[roll] Roll fuchsia [vfs] Use fidl sync binding in remote_dir for thread safety
The vfs::RemoteDir type holds a client binding to the
fuchsia::io::Directory interface which it uses from multiple threads.
This changes that binding to use the fidl::SynchronousInterfacePtr type
instead of fidl::InterfacePtr as the latter is thread-hostile. As this
type only makes calls to a FIDL method that has no response (Clone), it
is unlikely that the current usage causes concrete issues but it's
better to stick to documented safe usages.

Also removes the unnecessary dispatcher parameter and an unused
constructor.

Test: Caught by current tests with additional instrumentation being
developed in https://fuchsia-review.googlesource.com/c/fuchsia/+/446941/
Original-Bug: 63593

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447196
Original-Revision: b7c26daba56eb157330deb1e04083f0574bc7087
GitOrigin-RevId: 7c322e6913820be7af900c93d472d5b4e4d21397
Change-Id: I3e331ba9dcf35121c6f5e2dc212e155f92fe7b53