[roll] Roll fuchsia [elflib] Add explicit LLVM dependency
elflib is only using the ELF data structures, which are defined in LLVM
headers. It's important to provide a header-only LLVM target, as elflib
is built on both target and host, while LLVM is only available on host.

An alternative way is to avoid LLVM and use elf.h, which is not
available on macOS and requires certain config for compatibility.

Original-Bug: 47403
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446282
Original-Revision: 90bf1b1381fb86142efe454b001a27968fc29596
GitOrigin-RevId: 3d89454325701b65d1e9817d2d1a10cc139c1d47
Change-Id: Ie507c547d5e220ab4165f0969052df3a5dbb81c4