[roll] Roll fuchsia [media][tools] Fix warnings in signal_generator
This CL fixes "implicit conversion leads to precision loss" warnings
in the signal_generator CLI utility, as mentioned in fxbug.dev/58162.
To the extent that this limits the size of a payload buffer (4GB) or
the theoretical maximum duration of a signal (292 yrs), this is now
code-commented, user-error-checked, and stated in the '--?' output.

Test: No behavior change; diagnostic CLI utility
BUG: 58162

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446597
Original-Revision: 6af3810529fe8b61d874c088e6e647bb7d3c439c
GitOrigin-RevId: 79f1c5d6097432f21be1d870c1286730f1e29f73
Change-Id: I76bcdfb36d6689ab70a57d57324952472d056448