[roll] Roll fuchsia [bt][avrcp] Ignore unknown feature bits
When a peer is discovered, the AVRCP SDP supported feature
bits have a number of unspecified bits, which are reserved
for additions.  These bits should be ignored, and the connection
should continue only using the features it is aware of.

Original-Fixed: 45303, 46699, 46700
Test: IOPT/CL/AVRCP-CT/SFC/BV-03-I, IOPT/CL/AVRCP-TG/SFC/BV-04-I
      added service_from_attributes_ignores_unknown_features

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447108
Original-Revision: e4daddfb819c522abdd500b19a83b5b197c56e35
GitOrigin-RevId: 2da914b85c59eec74a1ce94ad2bc7e0bb5d8d252
Change-Id: I8f8ca45c0a56d2faa6908432afcd16831af72ce5