[roll] Roll fuchsia [tests] Use console.shell output for shell start
The QEMU tests used to rely on a virtcon message to know when
the serial console was up and running. Virtcon isn't responsible
for the serial console, the line that it should be looking for
is:
   console.shell: enabled

This CL was reverted earlier to see if it helped fix a flake.
The flaking test has since been disabled and is being
rewritten.

Original-Bug: 61528
TEST: Run the QEMU tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447456
Original-Revision: 3a79b5b2e7e27774115707a855dd045c00ec11f8
GitOrigin-RevId: d80476da837e247d55563bc8cb4096b6c168e4db
Change-Id: If1c606e34b368d87527e5db3cf0712bf3ba13e99