[roll] Roll fuchsia [kernel][vm] Annotate locked state
This annotates the VmMapping state to indicate the locks needed to
manipulate it. Due to limitations in the clang capability analysis it
cannot determine that if two locks are needed to modify a member, then
holding either lock makes it safe to read. Therefore multiple accessor
functions are added for the two lock options.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447058
Original-Revision: 18b931693a2b9e71ee0fc08433b46a53d47833d2
GitOrigin-RevId: 83212b308b79fcd7341c73406778e00cef5677c1
Change-Id: Id81e0caf5ebc352362c8b7bd175852dc31c28aa0