[roll] Roll fuchsia [kernel][vm] Transform into hidden parent
This changes the VmCowPages::CreateClone behavior to instead of
creating a new hidden VmCowPages and moving state into it, convert
the existing VmCowPages to be hidden and create new children that
reference it.

The benefit to transforming the existing node is that moving the page
list between VmCowPages is expensive, as every page has to be walked to
update any backlinks. The tradeoff here is that all the children have
to be updated to have a new parent, but children are expected to be
either 0 or 2, which is much less than the number of pages that are
expected.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437350
Original-Revision: f6fc82d67c1c477cf0cca0413748f41be39d0bfd
GitOrigin-RevId: af4d6d9d9a6301bf3467525bcaea6b50fd890ebe
Change-Id: Id76c5f9d24ee26f55df4242c4e0aba8591f61c4f