[roll] Roll fuchsia [rust][fidl] Fix lifetime issue on array of tables
This CL changes a test in src/sys/pkg/bin/omaha-client/src/fidl.rs to
explicitly bind an array of FIDL tables to a variable before calling
iter() on it. Although the code compiles as is, without this change it
fails to compile when the `{}` are replaced by `::empty()`. This will be
done soon as part of enforcing source-compatible tables in Rust.

Original-Bug: 7983
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447577
Original-Revision: 5c5eb3ba134d0b04fd96e575698f5758562dc313
GitOrigin-RevId: b88dade233424230b0487245b2a3d8b2e8e69522
Change-Id: Iaa69bdc3db8fc692264984aaf7f89f9a51c6350f