[roll] Roll fuchsia [pkg-cache] Maximize index iterator vector size
FIDL messages must fit within 64KiB, but it was possible for 32 index
entries to overflow a message (31 could not). This change increases the
maximum vector size for the package index iterator and modifies
pkg-cache to do the work to measure the size of elements in the
iterator, allowing it to fit as many items in a single FIDL message as
possible without overflowing it.

This change also fixes up the integration test for this iterator to not
need to construct a package for each index entry.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447194
Original-Revision: b5bfb953409933395b2b89403c25fd6c9716ba6f
GitOrigin-RevId: b6c1afc238927632ba16c1db4281c7ba9cf029df
Change-Id: I41d675390e045a68936f6177f3650c965fa44b14