[roll] Roll fuchsia [modular] Fix AgentSessionRestartTest racey use-after-free.
AgentSessionRestartTest sets up a session-agent and then issues a
request to restart the shell. The test waits then waits for the
session shell to stop, and restart, but does not wait for the
session agent to be restarted.

If the session agent restart request were processed after the
session has restarted, then the test will have completed and the
FakeAgent delete, causing the test to use-after-free.

Thanks to the ASAN builder for making this easy to diagnose!

Multiply: agent_session_restart_test
Test: agent_session_restart_test
Original-Bug: 63736
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447498
Fuchsia-Auto-Submit: Wez . <wez@google.com>
Original-Revision: 69adc7eb448967b169a312ebef44eeac12350353
GitOrigin-RevId: 87c5eed98213d862ca25353819c57b7e53d11a84
Change-Id: Id68321fafa322006ae86ce22f86b3e81f2e6adf6