[roll] Roll fuchsia [fvm] Change Metadata to store just one copy
Simplifies fvm::Metadata to only store one copy of the metadata (rather
than both of the A/B copies).

Because we are no longer planning to add snapshot information in a new
fixed-offset table (instead, we'll use reserved slices), it is no longer
necessary to hide the general layout of the metadata in the
fvm::Metadata object; instead clients can ask what offset/length to
write the A and B copies to.

This is simpler for when we integrate fvm::Metadata into the driver,
since the driver currently uses this model (where there is one in-memory
copy that is written to alternating offsets).

Original-Bug: 60709
Tested: Built and booted a variety of fvm images
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438495
Original-Revision: 1b412a7de1d44962445557605261f68b84515d3c
GitOrigin-RevId: 18f9f9022ad45312466c8efb796edd1fe9e2583b
Change-Id: I840b0731ce970bd992dd35c5a1dc74c11ec43143