[roll] Roll fuchsia [bt][rfcomm] Clean up credit flow mechanics
Updates the policy for including credits in frames. Based on behavior of
devices in the field, it's common to omit the credit octet when
the credits-to-be-issued is 0.

The Modem Status command is exchanged before user data is sent over the
DLC. GSM 7.10 Section 5.4.6.3.7 doesn't provide default values.
This CL updates the default value for the Data Valid (DV) bit to be set
based off of the behavior of devices seen in the field.

Moves channel.rs into the sessions mod as the SessionChannel is managed
by the RFCOMM Session.

Original-Bug: 63415
Test: Updated bt-rfcomm-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446277
Original-Revision: 496a67ba25a6a30f1f87fc119eb712207282ccfb
GitOrigin-RevId: 42fa6a92f0902926069023a6d5a117bf98112fb9
Change-Id: I8b61fca79133f8b7f779ffa43afd50dcff45b875