[roll] Roll fuchsia [rfc] Edge triggered async_wait
Documents a proposed change to zx_object_wait_async to
add a flag ZX_WAIT_ASYNC_EDGE which skips a check for
currently active signals and only queues a packet on the
port if a signal subsequently transitions from inactive
to active.

If any of the signals are active at the time of the
call of zx_object_wait_async, they must first transition
to the inactive state and then be asserted to generate
a packet on the port.

TEST: documentation only change.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442834
Fuchsia-Auto-Submit: Steve Austin <steveaustin@google.com>
Original-Revision: 0662651e1eeb5f25d8a4c5e22842c6c807133591
GitOrigin-RevId: 3ae0b5eaee0c24ce637cb3d2b9f3bfdc18d8c336
Change-Id: I9000c36363d814e9989528048369fedd8deba0f4