[roll] Roll fuchsia [fit] Move thread_checker functionality from fxl to fit
thread_checker is a type that can introduce runtime checks that a
sequence of calls are made from the same thread. This is useful for
adding runtime safety checks to thread-hostile types. This moves an
implementation of this logic from the internal fxl library to libfit so
that it can be exposed in the SDK and used by other C++ code published
in the SDK. In particular, some of the thread-hostile types in the fidl
library such as fidl::InterfacePtr could use this type in the future.

This change also renames the type and member functions of this type to
match the fit library style and updates it to depend on std::thread
instead of pthreads.

Test: Unit tests moved from fxl unit tests to fit unit tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446935
Original-Revision: d45b2a066b00edb0f6518b70f7b8db46e34205fd
GitOrigin-RevId: a1e7bac3f42b5b60569441ee10a98f2871c2a9e8
Change-Id: I5ae2c9568a779b6d4a94dede12cd99b2fb446c28