[roll] Roll fuchsia [test_spec] default to environments = basic_envs
By default (no gn args set) this is the same as the current default
of `[ emu_env ]`. The behavior will change for the case when the
gn arg `basic_names` is set.

We need this so that we can start running bringup tests based on the
contents of tests.json, rather than on specifying device_type in the
infrastructure. With this change, setting `basic_env_names` on a builder
will have a similar effect to the old setting of device_type.

Original-Bug: 59963
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447458
Fuchsia-Auto-Submit: Gary Miguel <garymm@google.com>
Original-Revision: b4858c46d625c0c9bd6b76a16adb1af828019e00
GitOrigin-RevId: 5a32675310558f1f4d05702502b1aa8ea594eb9e
Change-Id: I52bc966a3e673243135982d6e65b71136ea06143