[roll] Roll fuchsia [netemul] Add default route with specified NIC
... when using netstack_cfg.

Creating a default route without a specified NIC delegates the task of
finding the outgoing interface to the netstack. If a stack had 2
interfaces connected to the same network, the stack may only use
one of the interfaces as the outgoing interface. This will cause issues
when a stack may want to send packets through either of the interfaces
but only 1 will have the default route through it.

This change always specifies the NIC so that the default route goes
through the interface being configured.

Note that this is a limitation of the fuchsia.net.stack/Stack fidl which
does not allow clients to specify the outgoing interface when a route
depends on a next-hop node.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447674
Original-Revision: 4568eedc65a6fdb68736e93eea98055bead80e9b
GitOrigin-RevId: f7c1984a0a6fa7ee1237a3f904d7c729347c548c
Change-Id: I046f0782c0dc15937151011856ebce69fb90feef