[roll] Roll fuchsia [kernel][lib/arch] Introduce FakeCpuidIo
In this change, we update lib/arch to use BootCpuidIo (re-templating
accordingly) and introduce a in-test replacement for it: FakeCpuidIo.
FakeCpuidIo provides a simple interface for stubbing in dummy data and
is valid for use in the kernel as well as on host.

Original-Bug: 61093
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446377
Original-Revision: 0afca8e47fdbcf5d1c25fa7923fb2c429a75394c
GitOrigin-RevId: 26417ccf69d4f01afbbd5f1ba6a0af738c6994e9
Change-Id: Iba6c90186cd763a6b9e8a96e631f35c46efd3609