[roll] Roll fuchsia [securemem][logspam] Lower severity on non-bug logs
On boot, sysmem attempts to communicate with the securemem TA to
configure and discover secure memory heaps. This process failing is not
necessarily a bug, such as when sysmem is used during Zedboot. The TA is
unavailable in zedboot, so the secure heap configuration will fail, but
those heaps are not needed for zedboot anyways. This CL lowers the
severity to a WARNING and tweaks some of the log statements to include
relevant info.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447540
Fuchsia-Auto-Submit: RJ Ascani <rjascani@google.com>
Original-Revision: 37ff3bd295610546b453c6a00881cadd3f38ba51
GitOrigin-RevId: f3bcc608482359a9a2ec7def6e353548c91240c4
Change-Id: I8d026f2c9959ba80f8918e9008d6783f83a837bc