[roll] Roll fuchsia [metrics] Create the fuchsia.metrics FIDL API.
This is a copy of the current Cobalt 1.1 API from fuchsia.cobalt:
sdk/fidl/fuchsia.cobalt/cobalt.fidl

See go/fuchsia-cobalt-1.1 for the original design of the API.
The Cobalt 1.1 API under fuchsia.cobalt will be removed in a
subsequent CL, and the Cobalt app will be modified to instead
implement this new API under fuchsia.metrics. (There are no
users of the Cobalt 1.1 API yet.)

The API is unchanged in this copy, as the Cobalt 1.1 API is
already very generic. Almost all implementation-specific data
is passed using numeric IDs that correlate to fields in the
Cobalt registry. Any other implementations of the API can
define their own interpretations for these numeric IDs.

All of the changes made to the FIDL file contents are to the
comments, making them describe the API generically. Any
Cobalt-specific comments are moved the description of the API
in README.md.

Original-Bug: 61841
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444460
Original-Revision: 2d0c039fe0173a55b946a4b85d0c6dc9ec912f17
GitOrigin-RevId: 51b183b2f4dc36632fbbbf72e934ea138063bc50
Change-Id: Id267b952db947bd641757ed3ef16d4cc0f679dee