[roll] Roll fuchsia [size_checker] minor refactor of map code
Uses the result of map lookup as the node to put information in
when computing file sizes.

This is posible because the maps in question store node pointers,
so we can use the pointer alias to modify the map entry.  It is not
always possible, but in this case it is, and it reduces the chattiness
of the code.

Original-Bug: 63289
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447918
Original-Revision: 0e97c945d11e4c1bf54a855226c04209725e9ea3
GitOrigin-RevId: bf8294618cc8318f7c465b3d3b1c6049a91fec86
Change-Id: I553eee62cc8cf3bd19aad849cc8e09ce80feebb6