[roll] Roll fuchsia [wlan] Remove RCPI and RSNI
The RCPI and RSNI parameters are unused and only serve as a point of
confusion when determining how signal strength and broadband noise
measurements are passed around the different layers. Also, RCPI was
not technically in line with IEEE Std 802.11-2016 9.4.2.38 and was
instead simply RSSI / 2.

The topic of whether to use RCPI and RSNI, which are in fact the
802.11 units for measuring signal strength and noise, will be
revisited at a later date.

Note that part of this change substitutes rx_dbm with rssi_dbm
everywhere since the latter is no longer required because there is
only one unit.

Original-Bug: 60147
Original-Fixed: 62334
Test: Revised tests which used the RCPI parameter in wlan-common-tests.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445920
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Original-Revision: 248311494dd23022c483fce23e56b77865377a87
GitOrigin-RevId: 0314657b9f77cc7a7d83a0f9e3675e2fae9f3fa1
Change-Id: Icda34e105749f00e8c7f25fbb9040f60875db025