[roll] Roll fuchsia [settings] Serve values from light hardware configuration in light controller
The light API supports an optional hardware configuration that can be
specified through a json file. If the configuration is not present, the
light controller reads the underlying hardware API and reuses the names
the hardware returns, with each individual light getting it's own
LightGroup.

When a hardware configuration is specified, the product can define a
different mapping from physical lights to light groups, allowing logical
groupings of lights and specifying different options for them, such as
persistence.

While light controller already read a configuration previously, it did
not yet use it and still relied on the hardware lights. This CL will use
the light structure and names specified in the light hardware
configuration if it exists, falling back to the hardware otherwise.
Since the configuration is only used for tests at the moment, this CL
will not affect the behavior of the light API yet.

Original-Fixed: 63554
Test: fx test setui_service_tests, locally specified an configuration
and verified it changed the output of the light API
MULTIPLY: setui_service_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447300
Original-Revision: 2c316b8daadd8764ed63d43a288bac6ed5e6013d
GitOrigin-RevId: 9f4e5295ddfe1b082ed8fe7c76386ccba5253466
Change-Id: I229f8c0a84ba2615bf30a5d98aabb5c3157db678