[roll] Roll fuchsia [storage] Remove barrier support from block driver.
This also fixes a shutdown race: once TxnEnd has been called and
pending_count_ is decremented, the instance of Server can be destroyed
so it wasn't safe to access data members after that.  Removing support
for barriers means that no members are now touched in TxnEnd after that
point.

This CL removes the use of atomics which were unlikely to help with
performance, and replaces them with a conventional mutex and condition.

There was no need for the additional message queue so that has been
removed; messages can be immediately queued with the lower layers.

Original-Fixed: 41161
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/448080
Original-Revision: 7945929d1ce836482deb4cf49cf389e68da41f79
GitOrigin-RevId: d4d2f7263453a4ef793c4ea079f10b07af3911f0
Change-Id: I745dc8d9b737132f271b5a74c204813412dcb2e8