[roll] Roll fuchsia [image_pipe] Bind and use InterfacePtr on thread that uses it
fidl::InterfacePtr is a thread-hostile type so it must be bound and
destroyed on the same thread that it is used on. This updates the
ImagePipeSurfaceAsync class which uses a fidl::InterfacePtr on a
background thread to follow this guidance.

Original-Bug: 63593
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447916
Original-Revision: db2234d9c3d667bfda3bef5344180bd8d9f242e7
GitOrigin-RevId: 81dc8eec212ca4083bd3434f313b0c2e0a4e461d
Change-Id: Ic806d22b5b7a30d3279284c4619aa4a9e67e1aa2