[roll] Roll fuchsia [kernel] Timer: Convert active_cpu_ / cancel_ to ktl::atomic
Timer structs use active_cpu_ and cancel_ fields to track where a timer
callback is running and whether or not it was canceled. A timer can be
canceled from a different CPU than one that it's running / scheduled to
run on; it can be canceled while it is executing / as it is about to
start. Therefore the active_cpu_ / cancel_ fields need to be safe for
concurrent access.

The fields were marked volatile and arch::DeviceMemoryBarrier was used
after stores. Convert the fields to ktl::atomic, but keep the
DeviceMemoryBarriers in place at this point.

Most accesses can be converted to memory_order_relaxed; the cancel_
field in particular is set to indicate a timer is canceled. No dependent
accesses happen based on the cancel_ flag, so we do not need stronger
orders.

Tested:
1. k ut -r 100000 timer on qemu-kvm / x86
2. k ut -r 10000 timer on qemu-kvm w/ kCSAN

Original-Bug: 47117 Use only ktl::atomic in the kernel
Original-Bug: 61451 Kernelmode Concurrency / Thread Sanitizer for Zircon

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444274
Original-Revision: bfadca359ef98d84e6def16036f8bae6467f3556
GitOrigin-RevId: 1547d67f45240fb49f725b39ff5a0b732d80d2fb
Change-Id: I319fd92272dee72d3a5b76c4fa64dc932c1a626b