[roll] Roll infra/recipes [testing] Show timed out tests as "timed out"
If a test failed and its duration was greater than or equal to the
"per_test_timeout_secs" set for the build, we'll mark is as "timed out"
in Milo instead of "failed". This will make it easier to determine
whether a test failed because of a timeout.

This isn't a perfect heuristic, but in practice I expect it'll work
pretty well until I get around to fxbug.dev/49266, which will change
testrunner to explicitly mark timed out tests as timeouts.

Original-Bug: 60838
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/448856
Original-Revision: a15f4abf5ced74c00823db16bd67c896207c586a
GitOrigin-RevId: 05b143f670ea1f4ae73691218f761df994d6166e
Change-Id: Ie37a7639d6665e959635399ccdce9b7dd4e039f4