[roll] Roll fuchsia [opaque_test] Add dependency on component_manager.cmx
OpaqueTest does not have a dependency on
//src/sys/component_manager:component-manager.

This means that it is possible to build a test and
have the test fail because the component_manager.cmx
component could not be found.

Also remove the unnecessary direct dependency on
component manager for blobfs stress tests.

This happened in fxbug.dev/63817

Original-Bug: 63817
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449616
Original-Revision: 65c822cdcb59dad821230c1b9fc2fad0d1cc9a9a
GitOrigin-RevId: ad523f90f84e833e6c69ef36169ce5f8ff46f3f3
Change-Id: If59740662767d5292c2a720a0700c182b12cfaee