[roll] Roll infra/recipes Reland "[testing] Show timed out tests as "timed out""
This is a reland of a15f4abf5ced74c00823db16bd67c896207c586a

The original change compared the test timeout, in seconds, to each
test's duration, in milliseconds, so if a test exceeded 0.001 of the
timeout then it was considered to have timed out. Instead, we need to
convert the timeout to be in milliseconds before comparing to a
millisecond duration.

Original change's description:
> [testing] Show timed out tests as "timed out"
>
> If a test failed and its duration was greater than or equal to the
> "per_test_timeout_secs" set for the build, we'll mark is as "timed out"
> in Milo instead of "failed". This will make it easier to determine
> whether a test failed because of a timeout.
>
> This isn't a perfect heuristic, but in practice I expect it'll work
> pretty well until I get around to fxbug.dev/49266, which will change
> testrunner to explicitly mark timed out tests as timeouts.
>
> Bug: 60838
> Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/448856

Original-Bug: 60838
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/449737
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Original-Revision: ff586636946ecb6e6eeb3eec9726d71fdf84499f
GitOrigin-RevId: 9dd30bb16db244e6154d0b985b942dfbcfca029d
Change-Id: I0f9da166140f18a3f36687dc390b3580db4ecf11