[roll] Roll fuchsia [session] Remove assert on exposed services from component_manager_sfw
This check needs to be updated if the session_manager cml manifest
changes (e.g. when exposing a new protocol), and couples
component_manager_sfw to specific protocols. It also runs on each
invocation, not just in tests.

It might be better implemented as an integration test in the future.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449380
Original-Revision: 847123f7c94693126011d995957adaa3d848c0bb
GitOrigin-RevId: bba15c31cb9a9dbb4f04acc8b727cca50d1854c9
Change-Id: I7014c115e80488d82c99c2d1f8ce9063c17386c2