[roll] Roll fuchsia [fdt] Update fx fdt to handle moving our linux binary
We previously did the work of making fx fdt handle the new binary name
(fuchsia_devtools instead of system_monitor) however because our code
was still building from the system_monitor directory in google3, the
directory structure in our prebuilt still has that name. The final piece
of that story is about to land in google3 where we are doing our final
move of the code. We need to handle both structures so we can push that
forward and then we can finally come back and cleanup the _OLD stuff.

As we have done the binary name transition already we can replace the
current _OLD stuff with the current not old stuff as that is no longer
needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449170
Original-Revision: 2d9c59d35685925a4324f09f8c25b2e7b98f3ebb
GitOrigin-RevId: 5ff3ff08f9efa2c731f20fc1933b7974ad02c74c
Change-Id: I640e526df83e0121cec51ffd5da1abbc23ad08b0