[roll] Roll fuchsia [fdt] Update fdevtools.sh to handle moving our linux binary
We previously did the work of making fdevtools.sh handle the new binary name
(fuchsia_devtools instead of system_monitor) however because our code
was still building from the system_monitor directory in google3, the
directory structure in our prebuilt still has that name. The final piece
of that story is about to land in google3 where we are doing our final
move of the code. We need to handle both structures so we can push that
forward and then we can finally come back and cleanup the _OLD stuff.

As we have done the binary name transition already we can replace the
current _OLD stuff with the current not old stuff as that is no longer
needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449171
Original-Revision: bb1c13af3e245c8adea87d22a21ab2082164793b
GitOrigin-RevId: 8ce4e2cffa5e8a3b844acca189e95d5c94849246
Change-Id: I943c1da23d29afa1f0b3d97331140161803a8411