[roll] Roll fuchsia [fuchsia-runtime] Panic if runtime handle is not available
zx_clock_read will not fail even if the clock has never been updated,
it instead returns the backstop if it was created with one, or 0
otherwise. This means that utc_time() cannot fail unless the process
was launched with no UTC clock or a handle missing the required rights.

The documentation for zx_clock_read is updated in fxr/448994.

Original-Bug: 61741

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449074
Original-Revision: 303e40b9f326c2c75f1688d826d8987ab0aee677
GitOrigin-RevId: 4ec2755ff4c9b645318dea634a0424c955ba27f0
Change-Id: Ie7f3377ddba672c67c29464d12de4a4e36a36fd1