[roll] Roll fuchsia [settings] Refactor media buttons agent to use a hashset of settings
This CL updates the media buttons agent to calculate the settings that
it wants to send events to and store it as a set, instead of using an if
statement for each setting. This is done in preparation for adding light
controller as another setting type to receive media button events and
makes it easier to add additional setting types in the future.

Original-Fixed: 64142
Test: fx test setui_service_tests
MULTIPLY: setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449717
Original-Revision: 3aa91edd92b73083859636b521403ddff7de48da
GitOrigin-RevId: ae074fc4f73408c1a2040d65b1e22e79394d3023
Change-Id: I3f14d36886d4a087ad14b2f7de1b4ce66b79d5c8