[roll] Roll fuchsia [testrunner] Don't fail for snapshot failures.
Usually a failure to get a snapshot is really an SSH failure or some
other test failure that caused the snapshot to fail as well.
https://fuchsia-review.googlesource.com/c/fuchsia/+/446298 took care of
returning the SSH failure over the snapshot failure but it doesn't take
care of the case where the failure is caused by a test failure. Since
the snapshot is used more for debugging purposes, we don't need to fail
the task if it fails to run. This way we can surface the actual failure
if it's caused by a test failure. In the case that all tests pass,
however, tefmocheck should find and return the snapshot error.

Original-Bug: 61436
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449799
Original-Revision: 5ab09d328b135a9f31d810cdbcbe1deac0940ee3
GitOrigin-RevId: 01bc13b2e51045acc1b63f4e3c33925ad4049dad
Change-Id: Ie0d1700325dd8ca0361df8aea74f99365c3e3f63