[roll] Roll fuchsia [timekeeper] Address cobalt unittest flake.
There are two potentially causes of a mismatch in the expected clock
difference, being prempted between the two calls to initialize the
test clock objects and being prempted while reading the clock values.

The latter could be a problem leading to inaccurate data in real life
(on an emulator the test is flaking with slightly over 1ms difference)
so this CL has a few attempts to get a close pair (if we were prempted
on the first attempt the second atempt should be right at the start of
the next timeslice so I wouldn't expect more than one failure in a
row). I've also increased the tolerance on the check to help handle a
premption when setting the test clock.

Multiply: timekeeper-tests
Original-Bug: 63184
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449534
Original-Revision: 2b0843ab3b65539f8e2de7e7ff3082cb60f8dc12
GitOrigin-RevId: 7cd18caa63a4520f1ba240ddc62f517b9b7097ae
Change-Id: I3b412d0769535f8178e72fcf0bf32720483e9e57