[roll] Roll fuchsia [ffx][daemon] Drop targets from `ffx target list`.
When targets are deemed "disconnected" they will now be dropped from
`ffx target list`

This is unit tested but was also tested manually against emulator
targets.

This adds:
-- An atomic state updater for the target that updates the
   connection state.
-- Events for state transitions (currently only used in tests but could be used
   in a target event handler of some kind as well).
-- An MDNS monitoring task that drops a target when the state is left in
   MDNS with no state change and no repsponses (plus a grace period).

The intent of the connection states are that RCS and Fastboot supercede
mDNS, as they are able to more reliably determine if the target is up or
not, and as such will transition directly to
ConnectionState::Disconnected.

Original-Bug: 62784

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446856
Original-Revision: 6da4287f2c338aab59364f6a373d53fc6f7ef27a
GitOrigin-RevId: 55155c1863452dbbd9b77dc876f6a774ab35d3c2
Change-Id: If3ffd920cc7582f9e35c901f0efef963573b35f8