[roll] Roll fuchsia [storage][blobfs]: Add ability to visit each blob.
VisitBlobs is the first step to exporting each blob in the image.

The goal is layering the parsing of each blob and the writing of them
to simplify testing.

The following CL, just wraps this method up and writes the blob
as a child of directory fd.

Updated test not to rely in the current broken behavior of blobfs host
tooling (single extent, node ordering, etc).

BUG: 58653

TEST=CQ

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449418
Original-Revision: d631cf40d9b220c3594369c24c82c3fa88eeec99
GitOrigin-RevId: ef8dc1f7814b8cbec5fcea9bb6fee806dce0caa2
Change-Id: I7371d5ce09aa7f2dc18a7e5c77be1b16c75df33f