[roll] Roll fuchsia [ota-dependency-checker] create ota-dependency-checker component
This change creates the v2 ota-dependency-checker component. In order to
save space, it shares a binary with the pkg-resolver.

The integration testing is done in v1 because we are still in the
process of figuring out how we want to do v2 integration tests. We'll
migrate the tests to v2 in fxbug.dev/64125.

Test: fx test -o ota-dependency-checker-tests
Test: fx test -o ota-dependency-checker-integration-tests
Test: fx test -o grand-swd-binary-tests
Original-Fixed: 63641
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449417
Original-Revision: f54147ddf8c51358c6380d35679cb839026592e4
GitOrigin-RevId: 92329c08f04c825fdd9f51f2f3fca06faf42e1d2
Change-Id: I25245d81c8515aa9dc99ef704efe219950fa64ea