[roll] Roll fuchsia [storage][blobfs]: Add ExportBlobs command.
Operates over a generate blobfs images, taking advantage of |VisitBlobs|
writes each blob as an independant file in an output dir.

Added test to verify behavaior. Previous CL that introduced VisitBlobs
verifies that the contents provided to the visitor match the source, and
here we verify that the exported blobs match the contents of the files
used to generate the image on a first place.

Original-Bug: 58653

TEST=blobfs-host, CQ

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449137
Original-Revision: 9c4e6335865449ead24d402d0d52c2555a069192
GitOrigin-RevId: 5346b62110a72a415e0df6c3a22c757ae72cd692
Change-Id: Ia7687e6c5362aaaa928701d4b5d7cad4c874aa57