[roll] Roll fuchsia [digest] Fix bug with padding compact trees
If a HashLists's list was a multiple of the node size then a debug
assert in Digest::Update would fail when trying to pad the node because
Finish had already been called on the Digest.  Allowing PadWithZeros to
called on a finished NodeDigest fixes this.

Original-Bug: 36663
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445469
Original-Revision: 9f61541fb0c7e4e19f5541047a07daeeafec69c5
GitOrigin-RevId: c7330c4c9ed1860613de3bd8992b4f738ef9d07b
Change-Id: Ib9c200c12f40748e940cdbf84304887e8e75a996