[roll] Roll fuchsia [nix] implement edge-triggered polling in epoll
This enables the EPOLLET flag in epoll_ctl. When ORed with
the op flag this will cause epoll_wait to return a file
descriptor only when the file descriptor transitions from
inactive to active.

When a file descriptor controlled by EPOLLET is returned
from epoll_wait, non-blocking I/O should be performed on
the descriptor until EWOULDBLOCK is returned before calling
epoll_wait again. If epoll_wait is called when the file
descriptor is still ready, it may not return the descriptor.

Note that, if I/O is performed until EWOULDBLOCK is returned,
but the file descriptor becomes ready before the next call to
epoll_wait, the descriptor is guaranteed to be returned
appropriately, because of the use of ZX_WAIT_ASYNC_EDGE in
zx_object_wait_async.

Test: fx test nix-unittest-package
Original-Bug: 45709
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438656
Original-Revision: 2892e19ba734969108fdec9d1d1da4cb3ae1a317
GitOrigin-RevId: 1bf83f5f3ab37c0b6d400e3fa06b22013830e617
Change-Id: Iadaaaa8b0c508fe1fecd3244b99903e5930c9c1d