[roll] Roll fuchsia [sdio] Add RunDiagnostics method for extra logging after client errors
There have been several instances of the brcmfmac firmware download
failing on Sherlock without triggering SDIO CRC errors. On the surface
they don't appear to be related to SDIO, but it would be helpful to have
more information on the conditions of the bus and the card when this
happens. Add a RunDiagnostics method to the SDIO protocol which can be
called when the client encounters issues such as this.

Original-Bug: fxbug.dev/64166
Test: fx test sdmmc-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449817
Original-Revision: ba3f9d5098e7fecdceb15b5f3d77ba52b02338fc
GitOrigin-RevId: 88426fcef47b6621d5a6fb68c8a11b7885489443
Change-Id: I14885dc436e58e4be6399e16bae738cdd6ff8852