[roll] Roll fuchsia [ulib/fs] wait for remote to close on error
During remote mounting, the host filesystem essentially takes ownership
of the remote filesystem, and similar to channel ownership, it is
responsible for tearing down the remote filesystem when an error occurs.
This tear down on error was asynchronous - it wouldn't wait for a
response before finishing the request. This would cause tests to
occationally flake if they tore down the test environment before the
async filesystem shutdown finished.

This change makes Mount and MountAndCreate block on the shut down of the
remote filesystem.

Before this change, `fx test fs-management-test -- -f
MountFailsWithoutAdminCase.MountFailsWithoutAdmin` would fail every few
hundred times on qemu, with this change it's gone for about 10k without
a failure.

Original-Bug: 61269
Test: fx test fs-management-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450294
Original-Revision: 555057f8e5725964cfca91954c90918450fa60b2
GitOrigin-RevId: 1fed323662a79ffbfccc20f2554f814b73b7e89a
Change-Id: I001135cbbf68fe54c12e1da4ae0f54f3e9139839