[roll] Roll fuchsia [kernel][test] Fix a flake in PI tests.
Fix an issue with basic PI testing which is prone to flake.  The test
code is attempting to do the following:

1) Create two threads, A and B, and one OwnedWaitQueue Q.
2) Block thread A on an event.
3) Block thread B with a timeout on Q while declaring A to be the
   owner of Q.
4) Observe the PI effects of B on A transmitted through Q.
5) Allow B to timeout.
6) Observe that the PI effects in step #4 go away.

The problem with this test is that it is technically possible for
thread B to timeout before the observing thread has a chance to
observe the new effective priority of A.

To help avoid this, we both increase the timeout from 50mSec to
200mSec, but we also verify that the effective priority that we
observe was observed while thread B was still blocked.  If it has
unblocked during the test, we try again instead of failing the test.

Original-Fixed: 63552
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/448614
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Original-Revision: 89b35922338c0ef9ecc10db123762392cee64402
GitOrigin-RevId: ac59e33baf83e26593add15f074f8cc1073184da
Change-Id: I2f7db509ae1abe28f3aea3f08e476bc32b2917e8