[roll] Roll fuchsia Revert "[budget] add budget info for use in binary-sizes"
This reverts commit 723b4c42f83aaea3cdc01238075a28b079221d4a.

Reason for revert:

In testing the support in the binary plugin, we overlooked the effect
of having budget info in both the try builder AND the ci builder
(which only happened after this merged). The result was skewed plugin
rendering as it created superfluous rows. This issue is resolved in

https://chromium-review.googlesource.com/c/infra/gerrit-plugins/binary-size/+/2537150

So we should be able to re-land this once the updated plugin is
released.

Original change's description:
> [budget] add budget info for use in binary-sizes
>
> As part of go/fxb/63619 we aim to add  support for
> optional budget info in the binary-size gerrit plugin.
>
> This change to the size_checker tool's output will
> provide that info for consumption by the plugin as part
> of the size info JSON output.
>
> Bug: fuchsia:63619
> Test: cd tools/size_checker/cmd; go test
>
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/448294

TBR=garymm@google.com,leetodd@google.com,yifeit@google.com,fsamuel@google.com

No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Original-Bug: fuchsia:63619
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450375
Original-Revision: 78d6437248e72df13e180592c156f57a88428d5e
GitOrigin-RevId: 5f11e68651bb71475f743020bec4501d988b88d9
Change-Id: Ieb559c92c46113be4cf9c5ba3483c37755c2d337