[roll] Roll fuchsia [zircon] Rename "__USED" with "__ALWAYS_EMIT".
The compiler attribute "__USED" indicates that a particular variable is
used and should be emitted, even if the compiler can't see any
references to it.

It is rare to need the attribute, but it can be easily confused for the
attribute "__UNUSED" (which indicates "don't warn if this variable or
function is unused").

Rename "__USED" to "__ALWAYS_EMIT" to reduce confusion.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449920
Fuchsia-Auto-Submit: David Greenaway <dgreenaway@google.com>
Original-Revision: 395ae2c5ad2957ffd1a865d81e75318e18f31469
GitOrigin-RevId: 3d3ebbf0e90ae401d9a38186b6907ec5d54ae2fd
Change-Id: Ic710b026c583c0001fae3e4b5c176d9b7e55205f