[roll] Roll fuchsia [fvm] Support partition limits.
Add support for limiting a FVM partition size. A maximum byte size can
be set for each partition inside FVM, and slice allocations beyond this
size will fail if it would cause the partition to be larger than that.

The maximum partition size is not persisted. It must be set each time
after boot. We plan to do this from fshsot according to some
configuration values. It defaults to "no maximum" for each partition
until explicitly set.

safemath is used in new code and in some existing place which clarifies
the overflow checking. Checked casts are used in a few loops to make it
clear it's guaranteed the value doesn't overflow (I don't think there
was an actual bug here) and the extend and the count is 0-checked in the
allocate and free functions (this probably isn't a bug either but the
allocation unwind code in the case WriteFvmLocked() relied on some
wraparound and casting details in the 0-case that looks suspicious).

Original-Bug: 64055
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449954
Original-Revision: 46a0dfdf706424aa51f20b86d3dd0ef8da158e1b
GitOrigin-RevId: 540dc3515aea5c7dd427bd2fd31a9cf4cdf1b97a
Change-Id: I6822da3c5a631bb6254d78c26d98b5461c66f61a