[roll] Roll fuchsia [ramdisk] Fix ramdisk sleep race.
There was a race where the sleep-after block count could be set just as
we were processing a write and it would cause the counts to underflow
which was the cause of a flaky test.

Also fixed a blobfs test which was making the wrong assumption about the
device block size.

Original-Fixed: 63846
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450318
Original-Revision: 015d5ea5eea47c14d624d6b025bb169b3386d14c
GitOrigin-RevId: dc233c38477f67048eb507c06a9b3aed0f50c02c
Change-Id: I174b837deb16add3847c5d13b38a559984b1fdb8