[roll] Roll fuchsia [zircon][tests] Provide package container for kernel-clocks test
The kernel-clocks unit test is configured to run in a bringup
configuration as a ZBI item. This test can also run in non-bringup
configurations.  This adds a unittest_package() rule for the
kernel-clocks unit test to allow it to also run in these configurations.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449831
Original-Revision: 902933edb1282bc2670fd556404712f076f05000
GitOrigin-RevId: d6ab13a3d21b6374618476bef4a1e82e3f8d3e39
Change-Id: Ic2a3f7162b515e02e165e6ba4a068099f3caab77