[roll] Roll fuchsia [kernel][vm] Cache attribution counts in VmMapping
Cache page attribution counts for the VMO range that is mapped by a
VmMapping, and return the cached value if the VMO and the mapping have
not changed.

Introduces a mapping generation count to track modifications to the
VmMapping which alter the VMO range that is mapped. The caching logic
uses the mapping generation count and the VMO's hierarchy generation
count to infer if an update to the cached page count is required.

The mapping generation count is updated from VmMapping::ProtectLocked()
and VmMapping::UnmapLocked().

Test: kernel-unittests

Original-Bug: 56779
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449555
Original-Revision: 42bf8e49f66bb0b9736947f22b3c71455ec01f9b
GitOrigin-RevId: 0b9c8f0831c78792d264201c721d2aa856b76f15
Change-Id: I688b112fa10b9b987c842b00ea92d02eb0820709