[roll] Roll fuchsia [netstack] Inspect per-socket info of closing TCP endpoints.
Currently, TCP endpoints are deleted from the internal endpoints map
on a socket close(). This is premature for TCP as the endpoints can
linger around for sometime after a socket close(). The closing
state-machine has multiple states where it could be waiting for a
timeout or an ACK from the peer. For debug purposes, it is useful to be
able to inspect the endpoint info and stats in such states.

This change updates the endpoints map only on receiving EventHUp from
gVisor stack. This also changes map to have a separate monotonic
counter as the key, thus avoiding usage of local handle value as key which
would not be relevant after closing the handle in close() now.

Original-Fixed: 62372
Test: fx test -v netstack_test -- -test.run=TestTCPEndpointMapClosing -test.count=1000
      fx test -v netstack_test -- -test.run=TestEndpointsMapKey
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/443875
Original-Revision: 36d4d35a895584d2cede89b20f3e7e0aa2de36ac
GitOrigin-RevId: c953c3a8b18eb328026936254ebab513514060db
Change-Id: I10eaad49c5b124dbeb37d8f326e63cbe91feeec0