[roll] Roll fuchsia [fvm] Reset partition limit for new partitions.
The partition limit is not persisted to the device, but the FVM driver
was persisting the limit in memory for the partition index.

This change resets the partition limit when a partition is created to
make sure there is no leftover data in the array.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450460
Original-Revision: ca1e0bcbd6869f08856521c9e43219323e677737
GitOrigin-RevId: 71c8471076dcdfe0bb3ffa60a8bd2419ea0aaba6
Change-Id: I63ad18ca177ccfc2a8b736b3a9b900cb02cf360f