[roll] Roll fuchsia [component_manager] Fix elf_runner tests
In the elf runner tests, an Echo message was received but the
responder oneshot channel was not unblocked. In the right scenario
the client would see that the oneshot was canceled and would crash
crash the test.

Fixed by resuming the client before ending the test.

Test: fx test elf-runner-test
Original-Bug: 61970
Multiply: elf-runner-test: 100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/451114
Original-Revision: 9228f8cd2baadbb0924277da8aff5db6dc9a3021
GitOrigin-RevId: 8f0b85d66135a0cd979dd90d37c1f2901d3b2233
Change-Id: I519d03d4327263097ab0bf4e379de41af83cb8a7