[roll] Roll fuchsia [nix] Add test case for close()ing an FD in an epoll set
This adds a test case for our epoll emulation that considers the case
of closing a file descriptor while it is in an epoll interest set. This
shows a difference in our current implementation compared to the Linux
behavior.

Original-Bug: 64296
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450305
Original-Revision: 8e976957a21a3b484bac13a7c471218b9844d79e
GitOrigin-RevId: 9089aa1b460647c7d9b97fbe7092ef404fdf8254
Change-Id: Ic8d9d018b206adcb94edb71d438cf4817565c456