[roll] Roll fuchsia [component_manager] Actions use SharedFuture
Actions represent an asynchronous operation on a component or realm
(such as stop, start, and delete). Actions are idempotent, in the sense
that two matching actions scheduled at the same time will only be
executed once.

Currently, this is implemented as a custom future. Change the
implementation to use SharedFuture, which is simpler.

A side effect of this change resulted in task ordering differences. This
caused some tests to fail which were inadvertently dependending on the
previous order (for example, assuming that when bind() returns, the
runner has actually started the component). Fix those tests to be
agnostic to the order.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450182
Original-Revision: a811c6a1055eff0b58489548924ebe701936d190
GitOrigin-RevId: 522c83d92e10fd835ce69123095c845a5b2ce8c9
Change-Id: Id48505e50c30ab5ded2325e27d81755b580dcac1