[roll] Roll fuchsia [fbl][docs] : Add fbl:: container docs.
Get started on writing a top level User's Guide for using the fbl
intrusive containers.

The idea here is to provide a conversational tone in an example heavy
doc which shows what the containers can do, and the various ways to do
the most common things.  Somewhere between a FAQ and a cookbook.

A comprehensive reference section is something which would be great to
have as well, but this is a start at least.

Bugs: 49748
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389453
Original-Revision: 46d5033661a09b15b93da8da0eca42b5b0f28acd
GitOrigin-RevId: 68ae4a52936cab6d84ac9994bdb16ebeaef4741f
Change-Id: I295da1e87add4548f55bc532c6da48fe904a5834