[roll] Roll fuchsia [zbitl] Make Write() imply EnsureCapacity()
We codify in the API (in documentation and in asserts) that writable
storage types must also implement EnsureCapacity(). This simplifies
logic in View and Image that current check for both as separate degrees
of freedom when, in practice, they aren't.

Original-Bug: 57244
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450755
Original-Revision: df9d754d474539511f27170b8636dc77d1cdc455
GitOrigin-RevId: 3f55dd8485326ad17ae55e0e1f0304b14a7e931e
Change-Id: I7b4362b9916a5990d1c2c100d97530f1d15bbf77