[roll] Roll fuchsia [fs] Fix flaky test.
The problem here was to do with the fact that we do not have precise
control over ordering and it was possible for one of the Flush calls in
FlushPending to be slow and come after a flush from sync, which would
result in the test seeing an extra flush.  The fix is to:

 + Change FlushPending so that it doesn't actually Flush if there is no
   pending work.

 + Remove the Sync from the test, which was wrong anyway; the test is
   supposed to make sure it works without Sync.

In addition to the above, I have taken the opportunity to remove
barrier_ which was superfluous; journal_sequencer_ can be used instead.

Original-Fixed: 63379
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449919
Original-Revision: 8c1d29d8e267686bb0fd98511dfd8b70520e4ec9
GitOrigin-RevId: ffb00a668d4aca5dedfb6d2f9803eda47d10fecb
Change-Id: Ib7f32b5b3e77edd0052267c3b991cd4c9c22b673