[roll] Roll fuchsia [a11y] Relands TTS callback change with fix.
This change enables clients to register a callback with the tts manager
to be executed when the tts engine and speaker are connected and ready
to produce output. The previous version of this change introduced a bug
in which the screen reader would register a self-referencing callback with the TTS
manager, but would fail to unregister the callback on destruction. As a
result, the TTS would invoke a stale callback when the screen reader was
later re-enabled, resulting in a crash. We have addressed this issue and added a UT for this
case, so this change is now safe to re-land.

BUG: 63889
TEST: fx test a11y_lib_tests; fx test a11y_tests; manual
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450635
Original-Revision: 143178e3bb163758f0ef2990a447a3a296a6c73e
GitOrigin-RevId: bbfeb06f97bf41b54a4c9b702a3d339696b59237
Change-Id: I92f1eb78ae8cc7a17a91496942e0d40760f50618