[roll] Roll infra/recipes [tricium] Skip running go vet on uncompilable packages
There's a lot of Go code in fuchsia.git that imports generated files
(e.g. FIDL and protobufs) that are not present in a raw checkout. The
easiest way to keep `go vet` working in Tricium without breaking on
these packages is to skip vetting a package if its code isn't
compilable. With the `-json` flag set, `go vet` will only return a
retcode of one if the code isn't compilable (if it's compilable but
there are warnings, it'll still return a zero retcode). So we can
silently skip packages on which `go vet` returns a retcode of one.

Original-Bug: 63735
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/449979
Original-Revision: 59ed237c6f1334a5053ec34e3812c154ef02a052
GitOrigin-RevId: 12f23fc92a14e7d7cc95325955add2aefcb1963f
Change-Id: I797b766d3b82033897a77945c6e2926179438967