[roll] Roll fuchsia [components][logs] Routing log end-to-end test
Create a test topology where a component tries to access a
capability it does not have access to. In the test topology we
run Archivist and a reader component in the topology checks to
verify Archivists logs contain an error message for the routing
failure which looks about as expected and is attributed to the
requesting component instead of component manager.

Original-Fixed: 63000

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447075
Original-Revision: be1e5ca8bd3449db7ac8ba072f022082823180da
GitOrigin-RevId: 7eb17cbc458ec15038b4a2217ee177c1e640fcc8
Change-Id: Ide869749251f4ff7577564288f18e965c4c38147