[roll] Roll fuchsia [build] Specify path to vendored python binary in exec_script
This updates the exec_script() invocations in this repository to
specify the full path to the vendored python binary to use instead of
relying on the PATH or global script_executable to find it. This also
means that the path to the script, along with any other input files,
needs to be passed to exec_script()'s file_dependencies parameter so
that the generated ninja file re-generates when these inputs change.

This also adds the -S flag to speed up Python initialization for these
scripts.

Original-Bug: 63458, 62771
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446534
Original-Revision: bdb32a4e8d6d8484c8dfbd300516c2cce8a2684a
GitOrigin-RevId: 250a7c0af69f1f790d6be231a4b8bed3b56982f9
Change-Id: I91759970af8287418e1ee586e9130aaa7b270b99