[roll] Roll fuchsia [omaha client] correct calculation of update_check_response_time metric
The update_check_response_time metric is supposed to represent the
reponse time of making a single request to omaha (both success and
failure).  It was not being captured in the failure case, and in the
success case, it was the total time spent waiting between retries,
as well as the total time spent waiting for a failed attempt to occur.

This switches it to a per-attempt metric, also capturing the success
or failure of the result (the metric already had a dimension for this).

Original-Bug: 60557
Original-Bug: 60447
Original-Bug: 60448
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447298
Original-Revision: 87a229613d8f8d94ba5424c571dc3a4dffdae910
GitOrigin-RevId: 9be9d6aab9ddd398d4cb060da31e50d4198129ca
Change-Id: I20c260cc0c17118b335b0978d72e5a8dd7b0bc57