[roll] Roll fuchsia [kernel][x86][lbr] Add LBR k commands
Now that the global CpuidIo cache in the kernel is enabled, we can
properly use arch::LbrStack. As a proof-of-concept, we add an `lbr`
subcommand to `k cpu` that can enable, disable, and dump the LBR stacks
associated with particular CPUs.

Test: the k commands work on my NUC and return "not supported" in QEMU.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446380
Original-Revision: 6e5f8513d360f47c84ef56462696f95bd0a2c655
GitOrigin-RevId: cf29bdd998c20ca00bdc2acdf3814e4f09415c00
Change-Id: I93a290cd9029c72744fa7559e576a0a7f09d4233