[roll] Roll fuchsia [scenic/escher] Injectable GpuAllocator.
Escher can now optionally be provided with an alternate allocator
as a constructor argument.  If none is provided, then a default
allocator will be instantiated.

The goal is to allow Escher clients to know the concrete type of the
the allocator, which is otherwise opaque (i.e. users of the default
allocators know it is a GpuAllocator, but not which subclass).  Since
the client may want to continue to reference the allocator, Escher
now holds the allocator with a shared_ptr instead of a unique_ptr.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449718
Original-Revision: 26823130416db649b3f81f0a95b3df57032785e9
GitOrigin-RevId: 4ae1181fcc70c2f17b36aeee81a282c11bd3e801
Change-Id: I1b031c35d52730dc2bd1a0485592a33d83b9afec