[roll] Roll fuchsia [zircon_benchmarks] Clone VMOs wth multiple mappings
This extends the CloneMap VMO benchmarks to map the VMO using multiple
mappings, instead of a single mapping. The purpose of this is to
quantify the hidden impacts of VMO operations when multiple contiguous
mappings are used versus a single mapping, which is otherwise
functionally equivalent.

Original-Bug: 42207

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/448575
Original-Revision: 2e766b56800c48d0040c52fe82696b5a0965b8ad
GitOrigin-RevId: c4af4a6ddb5ea7027b9846f248fcb4af9279ce10
Change-Id: Ie553dd41fcd7fedac5e612b0fafd9f072613edf5