[roll] Roll fuchsia [fidlc] Support handle parametrization in aliases
Fixes handling of handle parametrization when used from a type alias.

We are purposefully not supporting partial type aliases since these are
being removed right now, as part of the implementation of FTP-052.

Interestingly, the `fuchsia.io2` library relied on this feature

    using Token = zx.handle:EVENT;

and the corrected handling in the compiler requires a source change. No
other FIDL file uses this sort of aliasing, and this change can land as
is without requiring an LSC.

Original-Bug: 64629
Original-Bug: 7807
Test: fx ninja -C out/default host_x64/fidl-compiler && \
      $FUCHSIA_DIR/out/default/host_x64/fidl-compiler --gtest_filter 'TypeAliasTests.*'
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/310005
Original-Revision: 090cabc293e38f5fe6708e73a805e352080ef20f
GitOrigin-RevId: 038f05d754a54f6bdc2c9e9194136b7317d1052f
Change-Id: I5379d0de4c385d2d3767cd837c546a13f45f7de4