[roll] Roll infra/recipes [tricium_analyze] Fix AttributeError and clean up
The exception described in the bug happened because of this line:

  ranges[i : i + 2] = Range(ranges[i].start, ranges[i + 1].end)

The intent of this line was to replace the items at indices i and i+1
with a single item. However, you can't assign a single item to a list
range in Python, only a list. So there should have been [] around the
Range(). However, this line actually didn't raise an exception because
`Range` is a namedtuple class, and namedtuples are iterable, so it's
valid to assign a namedtuple to a range of a list. So this line replaces
index i with the range.start and index i+1 with range.end. These are
integers, whereas the list is assumed to only contain Ranges, so when
the code later tries to access `start` and `end` of the integers it gets
an AttributeError.

I also did some general cleanup of the code to make it more
understandable and less error-prone, and I replaced the existing mock
diff data with data corresponding to the diffs that caused the failures
in production. This also gives us better code coverage, so I was able to
get rid of some `pragma: no cover` comments.

Long-term this code is way too complicated to be in recipes; it needs
actual unit tests. Probably not worth moving it into a Go tool now, but
it'll certainly benefit from a translation of recipes to Go if and when
that happens.

Before: https://ci.chromium.org/b/8863435559996080672
After: https://ci.chromium.org/swarming/task/4ff176ba3e489a10

Original-Bug: 63399
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/451428
Original-Revision: 3bc1e52f20db4dd2b5ae6acd002dd1651f0717ff
GitOrigin-RevId: 28a4edd3a5aa5920c1e616d043710a1dff0c4054
Change-Id: Ibbebc7370d2d153a98d5e48f45297740036c2cbf