[roll] Roll fuchsia [weave] Reduce the severity of logging.
Whenever a socket is closed, the corresponding fd waiter is cleared.
If no fd waiter is associated with the socket fd, it prints an error.
This is not a fatal error. So reducing the severity of the log.

Original-Bug: 64405
Test: Verified locally that logs are now at proper severity.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/451514
Original-Revision: 0c8f1d693c11c870e76de26a06671c178207a463
GitOrigin-RevId: 68002a23ea95b86ff8a4d9aec5796b4dc280e8eb
Change-Id: Ifebb0a8cbeb91382e3c1946382dc4752c14efe01