[roll] Roll fuchsia [fx][target] Order addresses by time. ULA support.
Adds ULA support when getting an SSH-able address for the target.

This takes target addresses and orders them by timestamp. This is to
make it possible for a target to maintain connectability upon several
reboots (as the target's address is no longer derived from its mac).

This also removes the TargetAddrFetcher trait as it was not being used
(just for some minor cleanup).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/451521
Original-Revision: 8fe2f52045c098c330ea61544d9cb3465b0ec2c8
GitOrigin-RevId: b4b51f996ce4f48882df158f05185879ead10a25
Change-Id: Ide2e00144f17cde1def6a5c02ab6fb1c2dd13fd8