[roll] Roll fuchsia [audio_core] Move client underflow/overflow logging to WARNING
When a client renderer underflows (does not queue packets quickly
enough), or a client capturer overflows (does not dequeue packets
quickly enough), we currently log that event to ERROR. This was done
because our FIDL APIs currently don't have a easy way for the client to
learn if they are underflowing / overflowing. Logging to ERROR seemed
better than nothing.

An ERROR log means "BUG". These logs are often evidence of bugs (e.g.,
fxbug.dev/50918), therefore ERROR is not entirely incorrect. However,
these logs are not always actionable as these are *client* bugs, not
audio_core bugs, and it's often difficult, if not impossible, to
correlate log entries with the offending client app. These logs also
occur in cases that are *not* bugs, such as when a streaming player
cannot download audio fast enough to keep up. The client can make this
distinction, but audio_core cannot.

Hence, since these logs are not always actionable, and not always bugs,
and since we get nags to "fix" these bugs (e.g., fxbug.dev/64589), this
CL changes the logs from ERROR to WARNING.

Original-Fixed: 53642
Original-Fixed: 64589
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/451939
Original-Revision: 564f30618974674f2c3dbc2154050996b089eb81
GitOrigin-RevId: a861bd8e6c14869e12f80c65660529ef78b0ea1d
Change-Id: I574f00f892bdd0267fc71c46a32f5bf9ff6069c3