[roll] Roll infra/recipes [tricium] Actually defer_results
I had a brain fart in fxrev.dev/452296 and put the
`defer_results()` in the wrong place. It needs to wrap the for loop over
all the changed files in the tricium.py reicpe.

Also add a test to confirm that Tricium comments are still emitted even
if an analysis step fails.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/452320
Original-Revision: 7a5b03e796a5250fcdc326770e3046ca356a8e21
GitOrigin-RevId: 85a0b321431153869fec6dd5780c3409048e5a08
Change-Id: I3595c4a8471480134b2d607a04b4d8a65beeb958