[roll] Roll fuchsia [lowpan][lowpan-spinel-driver] Initial Netstack integration
This commit provides the initial netstack integration with
lowpan-spinel-driver, using the `tun` interface for transporting
IPv6 traffic to and from the stack.

This is not the full implementation. The following features will be
provided in future commits:

 * Monitoring IPv6 addresses from netstack and handling their addition
   to and removal from OpenThread.
 * Monitoring IPv6 routes from netstack and handling their addition
   to and removal from OpenThread.

Additionally, the path the IPv6 packets takes through the driver is
sub-optimal from a performance perspective, involving several memory
allocations. This will be addressed later.

An integration test that makes sure that the network interface is
brought online properly is provided by <http://fxrev.dev/448839>.

Test: fx test lowpan-tests # + manual verification with ping
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434654
Original-Revision: 237f150bb5cb329ff40912ebf3282aa516c26c67
GitOrigin-RevId: 72d94a9cae083c8b3ef5e5d2fdab0c6d7c869e7b
Change-Id: I9fbcd021b93635eefb441ee844d7cc4ad8cb6b39