[roll] Roll fuchsia [blobfs] Set up client for external decompressor.
A thread safe client object that can be reused and a simpler wrapper for
the decompressor class that takes the client object. The client object
should attempt to set up the connection and fifo upon creation and upon
any loss of connectivity should recreate the connection on the next
attempted use.

Adds small integration test for the client with the component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446259
Original-Revision: 009e841b4bef24e25efd3ed4bedcf07f969eee0b
GitOrigin-RevId: 40e2121285b1c5970edbfc0e6f852a87e79d7879
Change-Id: I42cd68f6bc70a311965b9eb68a4b56edaf2d8d2a