[roll] Roll infra/recipes [build] Delete mock tool_paths.json data
See comment in build/api.py for rationale. Basically, it's a pain to
require that every tool that the recipe uses be in the mock
tool_paths.json, so we relax that requirement by returning a fake path
for any tools not in tool_paths.json when we're in unit testing mode.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/453218
Original-Revision: fa572aa5ad16817027555cfb151509d1e08ed8e0
GitOrigin-RevId: b79a216a48b30200ad3e2acc03e71e3c4356ac36
Change-Id: I9391ac5d49feec0416f10475fd25a71dc89cfceb