[roll] Roll infra/recipes [recipes][mac] Make failure to converge a warning for certain Mac paths
In some cases (see linked bug) checking relative paths on macOS will
occasionally return ENOENT (when the file actually exists). This causes
build problems including ninja not converging. It appears to be related
to traversing from the writable partition to the new read-only partition
introduced in Catalina.

Add a workaround to the re-run of ninja on Mac to ignore failures in a
few known paths. This is intentionally not a general suppression of
convergence checking to avoid regressing other no-ops and applies only
to Mac builds.

Original-Bug: 61784
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/450556
Original-Revision: 5148dcc136344725c143de32eec565d3030b7db6
GitOrigin-RevId: 7cbbac747cba7ed4ec63eb8579acad6d8942075a
Change-Id: I1c764afd1cb21c8079198a78a39bf9bc4b03f7ff