[roll] Roll fuchsia [weavestack] Shutdown on disconnect from netstack
When the FIDL connection to netstack is dropped, attempting to act on
sockets will return EPIPE (the underlying zx_channel is lost). In this
state, weavestack will continue to fail as it attempts to recover the
socket and will need to reconnect to the components it relies on. To aid
this, when the FIDL loss is detected, a task is now posted to the
application to request a shutdown of the main run loop.

A subsequent restart of the stack will have another chance to bind to
netstack and continue. This additionally prevents excess log spam from
weavestack when the system is in this state.

Original-Bug: 64223
Test: fx test weavestack-unittests
Test: fx test weavestack-adaptation-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/451530
Original-Revision: 6d2eb1771f06ba9bcef47c1afc2838c7291c04bb
GitOrigin-RevId: 50188c419d41b4f63761b9e6f9cc827c596822e5
Change-Id: I39438464fdc456984c452884b09154fd60ed6251