[roll] Roll fuchsia [bt][rfcomm] Clean up disconnection logic
Updates the clean up logic in the Session when a peer disconnects.

Context:
The RFCOMM Server maintains a map of Sessions indexed by the ID
of the connected peer. This map is updated lazily, meaning
stale Sessions will still exist in the map (even if the peer has already
disconnected). This is not an issue as entries are cleared and
reinitialized when the peer reconnects. However, multiplexed RFCOMM
channels will still remain opened (and clients none the wiser about peer
disconnection). Furthermore, component inspection will also reveal these
stale Sessions and can be misleading to debug.

This was not an issue pre- fxr/438439 because the SessionInner was not
shared, so when the peer disconnects the inner would be dropped (and
therefore cleaned up appropriately).

In order to maintain the proper RAII mechanics, this CL updates Session
to provide a termination signal, which can be polled by clients. The
RFCOMM Server uses this signal to eagerly evict the Session.

Test: Updated bt-rfcomm-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/451378
Original-Revision: a5f100c97ef830ed3905cf297a0fc05b5b0190e5
GitOrigin-RevId: 2ecabe7f75ec8e6ffb40c65c28d0e0014de99671
Change-Id: I4d84b1f09085e95bbf612c9c96aac61119fd5700