[roll] Roll fuchsia [mexec] Introduce //src/bringup/lib/mexec
This change introduces //src/bringup/lib/mexec, which is planned to be
the new site for userspace mexec logic. The `mexec_zbi_test()` template
will be moved to a testing/ subdirectory in a follow-up.

There are currently two places that benefit from this shared code, both
of which are updated here to make use of it: netsvc and the mexec ZBI
test chain-loading logic.

With this change we deprecate the fuchsia.kernel.MexecBroker protocol,
which no longer has any users (it will be removed in a follow-up). The
only benefit originally to the introduction of a FIDL service seems to
have been (a) abstracting the propagation of the root resource and (b)
language agnosticism. Regarding (a), the plan of record is now to
propogate an "mexec" resource down to any site that wishes to mexec,
and so there isn't a priori a need to propagate a second service down
to make use of that (a library call can do that with far less
overhead); regarding (b), both mexec callsites are written in C++ and
we do not wish to proliferate a convenient means to mexec anyway.

Test: the mexec ZBI tests continue to run successfully (and both the
x64 + arm devices on my desk - as well as those in CQ - boot happily).

Original-Bug: 57244
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/448115
Original-Revision: 2294350e9b78da502eee1020ad93d23743f2e9ff
GitOrigin-RevId: 361e049f8ee83009b6ec7bb26326321d8bf17944
Change-Id: I47c556cbf6e648c5125a6f9f0f8e7bb0951e292e