[roll] Roll fuchsia [devices] Add support for offers and exposed_dir
Until we have dynamic capability routing, we need a way to offer
services between drivers. This allows us to achieve that.

In the future, we will be able to remove exposed_dir, and capabilities
will be offer directly through the incoming namespace.

This is part of the drivers-as-components work.

Original-Bug: 33183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453156
Original-Revision: b3f1dbe491801b5ad4158b65245439840cf77a49
GitOrigin-RevId: 21fdd154fd044fd534b0a2014603322337c09006
Change-Id: I441cc5e16c23bef072ef5f5be20f629ed2f3741d