[roll] Roll fuchsia [kernel][arm64] Correct ASID flushing
ASID flushing in aspace destruction is presently incorrect as the ASID
should be in bits 63:48 of the TLB operation input register. This has
not caused problems as aspaces are presently fully unmapped (including
TLB flushing) prior to destruction.

This both corrects the ASID flushing with an additional wrapper macro
and moves ASID flushing into a helper function for the aspace.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453455
Original-Revision: 613b41073c105b202fc0593f564824b051e3c105
GitOrigin-RevId: 89d2b4a4a3bc2ab43e667443f3240c8472154feb
Change-Id: Ide080f565936f5bf69f909fd32e3ec564f856cb2