[roll] Roll fuchsia [audio_core] Avoid pipeline underflow on startup
Problem: when starting up a device, we perform some audio effect
inintialization. For effects that may take 'a long time' to initialize,
this causes the audio device to get behind on some of the initial frames
it needs to produce for the ring buffer. This in turn may result in the
first mix job needed to mix more than the expected 10ms, and combining
with the deadline profile this can lead to an pipeline underflow.

Solution: include the current driver safe write-ptr when computing what
inital frame should be sent. This results in us trimming those initial
frames (not mixing) instead of attemting to do the work to mix that
backlog.

Original-Bug: 62830
Original-Bug: 64368
Multiply: audio-core-api-pipeline-tests
Multiply: audio-core-google-pipeline-tests-astro
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425663
Original-Revision: a03d58fc017929fbcadf7744a9406c2c5c91fa1f
GitOrigin-RevId: 673716292306407c662e65c062da8ae22e122247
Change-Id: I789e19bdc145c443554eba00b088067e8f4858fb