[roll] Roll fuchsia [aml-rawnand] lower ECC errors to warnings
On raw NAND, ECC errors aren't always critical - we might be just
discovering that a block has gone bad, or we might need to retry a read
operation.

Lowering these logs to WARNINGS allows the higher-level caller to handle
it if possible, or log an ERROR itself if it really is a critical
failure.

Original-Bug: 64858, 64862
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453814
Original-Revision: 7d56a23bcb726a50eae5c71d8921dae281f17ab1
GitOrigin-RevId: 5d75f0f3c61cd8d46f2c3d269cae0c317ea8a4c1
Change-Id: Ibdd8fa24b8d46a985a32884f0dd4e75b0aa6dbe7