[roll] Roll fuchsia [media][audio_core] API fuzzer
A fuzzer is created to stress test the AudioRenderer and AudioCapturer
API calls in debug and asan modes. The fuzzer is built with an
understanding of all API method order dependencies, which enables it to
test edge cases along the understood happy path.

Original-Bug: 59621
Test: fx fuzz audio-fuzzers/api-fuzzer

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446482
Original-Revision: 9558da5919c45992744f1f98f486ea77caedab2e
GitOrigin-RevId: 9bd8e71e14546c2dd59427256563dee8e8298f76
Change-Id: Id251bb13697352a55d8582e447e55166dfad201d