[roll] Roll fuchsia [inspect][derive] Skip all fields except forwarded
When inspect is delegated to a child, skip all other fields on that
struct.

This makes forwarding match more closely documentation, and prevents
the requirement of a bunch of #[inspect(skip)] attributes on structs
that forward inspect management.

Make specifying any attributes on a struct where inspect is forwarded
into an error, and update to remove where this would break the build
because al other fields are skipped.

Update documentation and examples to make this behavior clear.

Test: updated tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450527
Original-Revision: a88b54126f3d51ee2d08f1c23db537c6870ac928
GitOrigin-RevId: 1b8284f20df745ea5189d9b52560bd31e4ba7041
Change-Id: I6f2384fd85e436a0eb287d3eaa15ee9d2aff31bf