[roll] Roll fuchsia [settings] Play max volume earcon on voice query
The earcons agent was only set up to play the max volume earcon
if the volume was increased via hw. This refactors so that sw changes to
max volume will also play the earcon if it is changed to max. This
eliminates the need for the media button events as the changes will all
come through the switchboard.

Original-Fixed: 42952
Test: fx test -o setui_service_tests
Multiply: setui_service_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453344
Original-Revision: 4aeef345cc672a0ff12afa4bf049d015471a2af4
GitOrigin-RevId: 87f39291099789c5d87b18f872b1a7317e4c00c2
Change-Id: Ib8ba110cb395c7273288c2fd31a00763e0e9b5eb