[roll] Roll fuchsia [components] Log open of non-existent path
If a component tries to open a path in a service directory in its
namespace which does not exist, log that attempt. Possibly the
component author forgot to include a `use` declaration. This is
an important debugging affordance whose true positives we think
pay for the false ones.

Original-Fixed: 64179

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453407
Original-Revision: 99b4078bd804143b2176345f4c4e682ed5c6a594
GitOrigin-RevId: a2750e5b2ddef68413c92f8139a8290b65dbb716
Change-Id: I01cf2769f122f3aca8f06c8b4e887236ea47b79b