[roll] Roll fuchsia [flatland] Kill channel on present with no tokens.
Flatland instances can kill the channel when the user makes an
unrecoverable error, such as presenting without any present tokens
available.

Ideally, the Flatland instance would just destroy the fidl::Binding
object, and the FlatlandManager would detect this in an async::Wait
and use that signal to cleanup the object. Unfotunately, the
FlatlandManager only has access to the server-side end of the channel
and cannot create an async::Wait to wait for that handle to close, so
it must instead pass a function into the Flatland instance that will
trigger destruction.

Original-Fixed: 62292
TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/452497
Original-Revision: 875781065a30aad05fc5d4139b60591b0232e543
GitOrigin-RevId: b5d5fdd7bccfbe68f106d52f1d0fb611da62f861
Change-Id: I3e8ca26928167a85b01ae7ac0bf6505e2d702541