[roll] Roll fuchsia [fidl][source-compatibility] Split up tests, generate docs
Split source compatibility tests so that each transition is defined in
its own library with its own independent source files. This makes it
simpler to reason about specific changes.

The steps of a transition are defined in a JSON file that's processed by
the build rules. The JSON file refers to each version of the FIDL
library (before / during / after) and then for each binding the steps to
execute a transition - either updating source code or updating the FIDL
library.

Based on the steps outlined in the configuration JSON each step of the
transition is built. Each version of the FIDL library has the same name
so that at compile-time each version of the source code can be paired
with the appropriate versions of the libraries.

This is an attempt to make it easier to scale source compatibilty tests
by splitting them into per-test directories with implementations for
each binding in that directory.

Documentation describing how to execute the transition based on the
steps outlined in the configuration JSON and the source code used in the
tests. The configuration JSON includes instructions describing the
changes that are made. These are rendered along side diffs of the source
and FIDL files into a markdown document.

Only one test has been converted so far. A replacement for the README.md
is in README.new.md once the old tests have been retired.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/439681
Original-Revision: d6ed250b0c00c4bca7a21f62082dbd5e75c662f2
GitOrigin-RevId: 84a001e3d3bd5a2103ccb17ad6553369028586e5
Change-Id: I464d003aa3c86fcd7a6b1db728b647aad76d05a1