[roll] Roll fuchsia [fx][emu] Fix terminal check script for "stty sane".
Previously we check if the shell contains option "-i" to determine
whether the script is running in interactive mode, and only reset
the terminal if script runs in interactive mode.

However when we run "fx emu", though the script is running in non-
interactive mode (because we spawn a new shell, instead of sourcing
the script in the interactive shell), the input and output are both
to a terminal; so the terminal still could be broken by QEMU, and
thus we still need a "stty sane" command after QEMU quits.

We fix this by changing the teriminal check script: we use "-t"
test operator to check if stdin is associated with a terminal
and only do stty sane when it is true.

TEST=fx emu (stty sane called)
     fx emu-remote (stty sane not called; no error messages).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453516
Original-Revision: eb18fbc3893495ad6c24815621cd30308423ebe5
GitOrigin-RevId: a7e3b684d7ef343d314a19e9ef789ff7c123002c
Change-Id: I9cb3d8736c000f7b2b74a2c3bf0607d9c97e32a9